pyproject.toml
[project]
name = "cryptolens"
version = "0.1.0"
description = "Near real-time Binance trade streaming pipeline to Apache Kafka"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka[avro]>=2.4.0",
    "websocket-client>=1.7.0",
    "pydantic-settings>=2.0",
    "fastavro>=1.9",
    "structlog>=24.0",
    "orjson>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cryptolens"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
